Better arg handling for properties/for-all
This specifically addresses the fact that in the old code failing
cases were reported as opaque functions, which is unhelpful.

Basically rewrote the whole thing by having the macro figure out
what names are bound exactly so that it can just generate a map of
symbols to pass into the property body.