Filtering out generated symbols
When trying to generate code to destructure the results of generators
with for-all, the macro was getting caught up by gen-syms. By filtering
out symbols with the prefixes "map__" and "vec__", the macro no longer
tries to generate code that references the generated symbols and works
as expected (or at least how I expected it to work). A simple test was
added as well.