Simply printout the result in case of failure
The checking macro was not printing the test.check result in case of
failures. This was happening since we changed the assertion on the
result of `checking` to only check if there was an exception. That
change was introduced to fix #17, which was about the output showing an
"expected: (not-falsey-or-exception? (:result result))" message. But
as a side-effect of that change, we lost the test.check result from the
output, which is very valuable.

The fix is to simply print the test.check result in case of failure.
This is part of what `do-report` would do if we called
`(is (:result result) result)` but now we avoid the other stuff that
call would do (incrementing test result counters, output the expected
vs actual result of the assertion, etc).

As a side-effect of this change, the `:pass` count in case of a failure
was fixed. That can be seen by the change in the failure-detecion-test